+++ pkg/ffuf/progress.go
@@ -8,5 +8,7 @@ type Progress struct {
 	StartedAt  time.Time
 	ReqCount   int
 	ReqTotal   int
+	QueuePos   int
+	QueueTotal int
 	ErrorCount int
 }